requirements.txt
wheel
numpy
pandas
nltk
matplotlib
networkx
seaborn

Pillow